Split form control components
