+++ package/src/helpers.js
@@ -1,0 +1,6 @@
+import React from 'react';
+
+export const withForwardedRef = Wrapped =>
+  React.forwardRef((props, ref) => {
+    return <Wrapped {...props} forwardedRef={ref} />;
+  });
