+++ package/src/input.js
@@ -1,189 +1,378 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import {withRadiumStarter, Input as OriginalRSInput} from 'radium-starter';
+import {withRadiumStarter, TextArea as RSTextArea} from 'radium-starter';
+import isEqual from 'lodash/isEqual';
+import omit from 'lodash/omit';
+import Downshift from 'downshift';
+import matchSorter from 'match-sorter';
 
-export const withForwardedRef = Wrapped =>
-  React.forwardRef((props, ref) => {
-    return <Wrapped {...props} forwardedRef={ref} />;
-  });
+import {withLocale} from './locale-context';
+import {withForwardedRef} from './helpers';
+import {RSInput} from './rs-input';
 
 @withForwardedRef
-export class RSInput extends React.Component {
+export class Input extends React.Component {
   static propTypes = {
-    type: PropTypes.string,
-    disabled: PropTypes.bool
+    value: PropTypes.string,
+    forwardedRef: PropTypes.object,
+    onChange: PropTypes.func,
+    onFocus: PropTypes.func,
+    onBlur: PropTypes.func,
+    required: PropTypes.bool,
+    tabIndex: PropTypes.string,
+    style: PropTypes.object
   };
 
-  getCustomComponent() {
-    const {type} = this.props;
-
-    if (type === 'checkbox') {
-      return CheckboxMark;
-    }
-    if (type === 'radio') {
-      return RadioMark;
-    }
-
-    return undefined;
+  shouldComponentUpdate(nextProps, _nextState) {
+    return (
+      nextProps.value !== this.props.value ||
+      nextProps.onChange !== this.props.onChange ||
+      nextProps.onFocus !== this.props.onFocus ||
+      nextProps.onBlur !== this.props.onBlur ||
+      nextProps.required !== this.props.required ||
+      nextProps.tabIndex !== this.props.tabIndex
+    );
   }
+
+  handleChange = event => {
+    this.props.onChange(event.target.value);
+  };
 
   render() {
     const {forwardedRef, ...props} = this.props;
-    const CustomComponent = this.getCustomComponent();
-
-    if (!CustomComponent) {
-      return <OriginalRSInput {...props} ref={forwardedRef} />;
-    }
 
     return (
-      <>
-        <CustomComponent {...props} style={{cursor: props.disabled ? 'not-allowed' : 'pointer'}} />
-        <OriginalRSInput
-          ref={forwardedRef}
-          {...props}
-          style={{
-            // Styles used to hide the underlying `<input>` tag.
-            // Don't use `display: 'none'` to be able to focus the checkbox with the keyboard
-            // Don't use `width: 0, height: 0`, otherwise HTML field validation messages don't show up
-            opacity: 0,
-            overflow: 'hidden',
-            margin: 0,
-            padding: 0,
-            position: 'absolute'
-          }}
-        />
-      </>
-    );
-  }
-}
-
-// sub-components (not exported)
-
-@withRadiumStarter
-class CheckboxMark extends React.Component {
-  static propTypes = {
-    checked: PropTypes.bool,
-    disabled: PropTypes.bool,
-    style: PropTypes.object,
-    theme: PropTypes.object.isRequired,
-    styles: PropTypes.object.isRequired
-  };
-
-  render() {
-    const {disabled, checked, style, theme: t, styles: s} = this.props;
-
-    const getBackgroundColor = () => {
-      if (disabled) {
-        return t.altBackgroundColor;
-      }
-
-      if (checked) {
-        return t.primaryColor;
-      }
-
-      return 'white';
-    };
-
-    const getTickColor = () => {
-      if (disabled) {
-        return t.altTextColor;
-      }
-      if (checked) {
-        return 'white';
-      }
-      return undefined;
-    };
-
-    const inlineSVG = encodeURIComponent(`
-      <svg width="11" height="10" viewBox="0 0 11 10" fill="none" xmlns="http://www.w3.org/2000/svg">
-        <path
-          fill-rule="evenodd"
-          clip-rule="evenodd"
-          d="M10.6 0.200059C11.0418 0.53143 11.1314 1.15823 10.8 1.60006L4.8 9.60006C4.62607 9.83197 4.36005 9.97699 4.07089 9.99754C3.78173 10.0181 3.49788 9.91215 3.29289 9.70717L0.292893 6.70717C-0.0976311 6.31664 -0.0976311 5.68348 0.292893 5.29295C0.683417 4.90243 1.31658 4.90243 1.70711 5.29295L3.89181 7.47765L9.2 0.400059C9.53137 -0.0417689 10.1582 -0.131312 10.6 0.200059Z"
-          fill="${getTickColor()}"
-        />
-      </svg>
-    `);
-
-    return (
-      <span
-        style={{
-          ...s.border,
-          ...s.rounded,
-          flex: '0 0 auto',
-          display: 'inline-block',
-          width: '20px',
-          height: '20px',
-          margin: '2px',
-          borderWidth: '2px',
-          borderColor: !checked && !disabled ? t.grayIconColor : 'transparent',
-          backgroundColor: getBackgroundColor(),
-          backgroundImage: checked ? `url('data:image/svg+xml,${inlineSVG}')` : null,
-          backgroundRepeat: 'no-repeat',
-          backgroundPosition: 'center',
-          transitionDuration: '0.3s',
-          ...style
-        }}
+      <RSInput
+        ref={forwardedRef}
+        {...props}
+        onChange={this.handleChange}
+        style={{display: 'block', width: '100%', ...this.props.style}}
       />
     );
   }
 }
 
-@withRadiumStarter
-class RadioMark extends React.Component {
+@withForwardedRef
+@withLocale
+export class TextInput extends React.Component {
   static propTypes = {
-    checked: PropTypes.bool,
-    disabled: PropTypes.bool,
-    style: PropTypes.object,
-    theme: PropTypes.object.isRequired
+    ...Input.propTypes,
+    locale: PropTypes.object.isRequired
+  };
+
+  handleChange = value => {
+    const {onChange, locale} = this.props;
+    if (onChange) {
+      onChange(locale.parseTextInput(value));
+    }
   };
 
   render() {
-    const {disabled, checked, style, theme: t} = this.props;
+    const {value, forwardedRef, onChange, locale, ...props} = this.props;
+    return (
+      <Input
+        ref={forwardedRef}
+        {...props}
+        value={locale.formatTextInput(value)}
+        onChange={this.handleChange}
+      />
+    );
+  }
+}
 
-    const getOuterColor = () => {
-      if (disabled) {
-        return t.altBackgroundColor;
-      }
-      if (checked) {
-        return t.primaryColor;
-      }
-      return t.grayIconColor;
-    };
+@withForwardedRef
+@withLocale
+export class NumberInput extends React.Component {
+  static propTypes = {
+    ...Input.propTypes,
+    value: PropTypes.number,
+    locale: PropTypes.object.isRequired
+  };
 
-    const getInnerColor = () => {
-      if (disabled) {
-        return checked ? t.altTextColor : t.altBackgroundColor;
-      }
-      return 'white';
-    };
+  state = {
+    draftValue: ''
+  };
+
+  static getDerivedStateFromProps({value, locale}) {
+    if (value !== undefined && isNaN(value)) {
+      return null;
+    }
+    const draftValue = locale.formatNumberInput(value);
+    return {draftValue};
+  }
+
+  handleChange = draftValue => {
+    const {onChange, locale} = this.props;
+    this.setState({draftValue});
+    if (onChange) {
+      const value = locale.parseNumberInput(draftValue);
+      onChange(value);
+    }
+  };
+
+  render() {
+    const {value, forwardedRef, locale, ...props} = this.props;
+    const {draftValue} = this.state;
+    return (
+      <Input
+        ref={forwardedRef}
+        {...props}
+        value={draftValue}
+        onChange={this.handleChange}
+        rsCustomValidity={
+          value !== undefined && isNaN(value) ? locale.numberInputInvalidityMessage : ''
+        }
+      />
+    );
+  }
+}
+
+@withForwardedRef
+export class TextArea extends React.Component {
+  static propTypes = {
+    value: PropTypes.string,
+    onChange: PropTypes.func.isRequired,
+    required: PropTypes.bool,
+    style: PropTypes.object
+  };
+
+  shouldComponentUpdate(nextProps, _nextState) {
+    return (
+      nextProps.value !== this.props.value ||
+      nextProps.onChange !== this.props.onChange ||
+      nextProps.required !== this.props.required
+    );
+  }
+
+  handleChange = event => {
+    this.props.onChange(event.target.value);
+  };
+
+  render() {
+    const {forwardedRef, ...props} = this.props;
 
     return (
-      <div
-        style={{
-          display: 'flex',
-          alignItems: 'center',
-          backgroundColor: getOuterColor(),
-          borderRadius: '50%',
-          width: '20px',
-          height: '20px',
-          justifyContent: 'center',
-          margin: '2px',
-          verticalAlign: 'middle',
-          flexShrink: 0,
-          ...style
-        }}
-      >
-        <div
-          style={{
-            backgroundColor: getInnerColor(),
-            borderRadius: '50%',
-            width: checked ? '6px' : '16px',
-            height: checked ? '6px' : '16px',
-            transitionDuration: '0.2s'
-          }}
-        />
-      </div>
+      <RSTextArea
+        ref={forwardedRef}
+        {...props}
+        onChange={this.handleChange}
+        style={{display: 'block', width: '100%', ...this.props.style}}
+      />
     );
+  }
+}
+
+@withForwardedRef
+export class CheckboxInput extends React.Component {
+  static propTypes = {
+    label: PropTypes.string,
+    value: PropTypes.bool,
+    onChange: PropTypes.func.isRequired,
+    required: PropTypes.bool,
+    style: PropTypes.object
+  };
+
+  id = String(Math.round(Math.random() * 1000000000));
+
+  shouldComponentUpdate(nextProps, _nextState) {
+    return (
+      nextProps.label !== this.props.label ||
+      nextProps.value !== this.props.value ||
+      nextProps.onChange !== this.props.onChange ||
+      nextProps.required !== this.props.required
+    );
+  }
+
+  handleChange = event => {
+    this.props.onChange(event.target.checked, event);
+  };
+
+  render() {
+    const {label, value, forwardedRef, required, style} = this.props;
+
+    return (
+      <label style={{display: 'flex', alignItems: 'center', ...style}} htmlFor={this.id}>
+        <RSInput
+          ref={forwardedRef}
+          {...omit(this.props, ['value'])}
+          id={this.id}
+          type="checkbox"
+          checked={value || false}
+          onChange={this.handleChange}
+        />
+        {label && (
+          <span style={{paddingLeft: '0.5rem', verticalAlign: 'middle'}}>
+            {label}
+            {required && <Asterisk />}
+          </span>
+        )}
+      </label>
+    );
+  }
+}
+
+@withForwardedRef
+@withRadiumStarter
+export class AutocompleteInput extends React.Component {
+  static propTypes = {
+    type: PropTypes.string,
+    forwardedRef: PropTypes.object,
+    id: PropTypes.string,
+    value: PropTypes.string.isRequired,
+    onChange: PropTypes.func.isRequired,
+    onFocus: PropTypes.func,
+    onBlur: PropTypes.func,
+    items: PropTypes.oneOfType([PropTypes.array, PropTypes.object]),
+    disabled: PropTypes.bool,
+    required: PropTypes.bool,
+    placeholder: PropTypes.string,
+    maxLength: PropTypes.number,
+    style: PropTypes.object,
+    theme: PropTypes.object.isRequired,
+    styles: PropTypes.object.isRequired
+  };
+
+  static defaultProps = {
+    items: []
+  };
+
+  state = {
+    items: []
+  };
+
+  shouldComponentUpdate(nextProps, nextState) {
+    return (
+      nextProps.value !== this.props.value ||
+      nextProps.onChange !== this.props.onChange ||
+      nextProps.onFocus !== this.props.onFocus ||
+      nextProps.onBlur !== this.props.onBlur ||
+      nextProps.disabled !== this.props.disabled ||
+      nextProps.required !== this.props.required ||
+      nextProps.items !== this.props.items ||
+      !isEqual(nextState.items, this.state.items)
+    );
+  }
+
+  componentDidUpdate({items: prevItems}) {
+    const {items} = this.props;
+
+    if (items === prevItems) {
+      return;
+    }
+
+    if (typeof items.then === 'function') {
+      // items is a Promise
+      items.then(items => {
+        this.setState({items});
+      });
+    } else {
+      this.setState({items});
+    }
+  }
+
+  handleChange = value => {
+    if (value !== this.props.value) {
+      this.props.onChange(value);
+    }
+  };
+
+  render() {
+    const {
+      type,
+      forwardedRef,
+      id,
+      value,
+      onFocus,
+      onBlur,
+      disabled,
+      required,
+      placeholder,
+      maxLength,
+      style,
+      theme: t,
+      styles: s
+    } = this.props;
+    const {items} = this.state;
+
+    let filteredItems = items.filter(item => item !== value);
+    filteredItems = matchSorter(filteredItems, value);
+
+    return (
+      <Downshift
+        onChange={this.handleChange}
+        inputValue={value}
+        onInputValueChange={this.handleChange}
+        selectedItem={value}
+      >
+        {({getInputProps, getItemProps, isOpen, highlightedIndex}) => (
+          <div style={{position: 'relative'}}>
+            <RSInput
+              {...getInputProps({
+                type,
+                ref: forwardedRef,
+                id,
+                onFocus,
+                onBlur,
+                disabled,
+                required,
+                placeholder,
+                maxLength,
+                autoComplete: 'nope',
+                style: {display: 'block', width: '100%', ...style}
+              })}
+            />
+            {filteredItems.length > 0 && isOpen && (
+              <div
+                style={{
+                  ...s.field,
+                  padding: '0.5rem 0',
+                  position: 'absolute',
+                  width: '100%',
+                  maxHeight: '190px',
+                  marginTop: '-1px',
+                  overflowY: 'auto',
+                  zIndex: 1,
+                  cursor: 'default',
+                  backgroundColor: t.inputBackgroundColor,
+                  borderColor: t.focusedInputBorderColor
+                }}
+              >
+                {filteredItems.map((item, index) => (
+                  <div
+                    {...getItemProps({item})}
+                    key={item}
+                    style={{
+                      padding: '0 0.625rem',
+                      color: highlightedIndex === index ? t.inverseTextColor : undefined,
+                      backgroundColor: highlightedIndex === index ? t.primaryColor : undefined
+                    }}
+                  >
+                    {item}
+                  </div>
+                ))}
+              </div>
+            )}
+          </div>
+        )}
+      </Downshift>
+    );
+  }
+}
+
+@withRadiumStarter
+export class Asterisk extends React.Component {
+  static propTypes = {
+    theme: PropTypes.object.isRequired,
+    styles: PropTypes.object.isRequired
+  };
+
+  shouldComponentUpdate(_nextProps, _nextState) {
+    return false;
+  }
+
+  render() {
+    const {theme: t} = this.props;
+
+    return <span style={{color: t.errorColor}}>*</span>;
   }
 }
